backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.4
python-dotenv==1.0.1
numpy==1.26.3
pandas==2.0.3
scikit-learn==1.6.1
drf-yasg>=1.21.0
psycopg2-binary>=2.9
requests
django-extensions
joblib>=1.3.0
coverage
gunicorn